Remove friendlies metatable
I personally don't see why __len and __index ought to be defined. It feels inconsistent with the rest of the API, and it's just more of the same. Also, it gives the illusion of `friendlies` (a "class", i guess) of being a mere object, without any notice of it changing between two Update()s.